pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-music-download"
version = "0.1.0"
description = "โปรแกรมดาวน์โหลดเพลงจาก YouTube และ YouTube Music ด้วย CustomTkinter GUI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = []
keywords = ["youtube", "music", "downloader", "yt-dlp", "mp3"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "yt-dlp>=2023.0.0",
    "requests>=2.28.0",
    "python-dotenv>=0.21.0",
    "customtkinter>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=5.0.0",
]

[project.scripts]
youtube-music-download = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
